A lot of refacctoring on the backend audio

+++ sources/Services/Audio/AudioOut.h
@@ -44,7 +44,7 @@ public:
    int getPlaySampleCount() ;
 
 private:
-	float sampleOffset_ ;
+	double sampleOffset_ ;
 
 } ;
 #endif
